+++ src/Order/Payment/PaymentComposer.php
@@ -1,9 +1,8 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 namespace Shopgate\Shopware\Order\Payment;
 
+use RuntimeException;
 use Shopgate\Shopware\System\Log\LoggerInterface;
 use ShopgateCartBase;
 use ShopgatePaymentMethod;
@@ -11,6 +10,7 @@ use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionColl
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionEntity;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStates;
 use Shopware\Core\Checkout\Payment\PaymentMethodEntity;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Shopware\Core\System\StateMachine\Aggregation\StateMachineState\StateMachineStateEntity;
 
@@ -65,6 +65,37 @@ class PaymentComposer
         $transaction = $this->getActualTransaction($transactions);
 
         return $transaction ? $this->paymentBridge->setOrderToPaid($transaction->getId(), $context) : null;
+    }
+
+    /**
+     * Retrieves customer payment method ID from context
+     * Defaults to channel if all fallbacks are inactive
+     *
+     * @param SalesChannelContext $context
+     * @return string
+     * @throws RuntimeException
+     */
+    public function getCustomerActivePaymentMethodId(SalesChannelContext $context): string
+    {
+        $methods = [$context->getPaymentMethod()->getId()]; // customer selected payment
+        if ($context->getCustomer()) {
+            if ($context->getCustomer()->getLastPaymentMethodId()) {
+                $methods[] = $context->getCustomer()->getLastPaymentMethodId(); // customer last order payment
+            }
+            if ($context->getCustomer()->getDefaultPaymentMethodId()) {
+                $methods[] = $context->getCustomer()->getDefaultPaymentMethodId(); // customer default
+            }
+        }
+        $methods[] = $context->getSalesChannel()->getPaymentMethodId(); // channel default payment
+        $ids = array_combine($methods, $methods);
+        $criteria = new Criteria($ids);
+        $activePayments = $this->paymentBridge->getAvailableMethods($context, $criteria);
+        $activePayments->sortByIdArray($ids);
+        if (!$payment = $activePayments->first()) {
+            throw new RuntimeException('Is SalesChannel default payment method not active?');
+        }
+
+        return $payment->getId();
     }
 
     private function getActualTransaction(?OrderTransactionCollection $transactions): ?OrderTransactionEntity